requirements.txt
flask
werkzeug
torch
torchvision
easyocr
mysql-connector-python
numpy
opencv-python-headless
pandas
pytest
gunicorn
ultralytics